# PERF: use NaT comparisons in int64/datetimelikes #11010