closes #11010 

```
In [10]: %timeit gr['3rd'].count()
100 loops, best of 3: 15.1 ms per loop

In [11]: %timeit gr['3rd'].min()
100 loops, best of 3: 8.4 ms per loop

In [12]: %timeit gr['3rd'].max()
100 loops, best of 3: 8.47 ms per loop

In [13]: %timeit gr['3rd'].first()
100 loops, best of 3: 10.5 ms per loop

In [14]: %timeit gr['3rd'].last()
100 loops, best of 3: 10.3 ms per loop
```